add remotecollect params struct
