+++ pkg/collect/host_os_info.go
@@ -2,19 +2,12 @@ package collect
 
 import (
 	"bytes"
-	"context"
 	"encoding/json"
-	"fmt"
-	"os"
 	"time"
 
 	"github.com/pkg/errors"
 	troubleshootv1beta2 "github.com/replicatedhq/troubleshoot/pkg/apis/troubleshoot/v1beta2"
 	osutils "github.com/shirou/gopsutil/v3/host"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/apiserver/pkg/storage/names"
-	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/rest"
 )
 
@@ -77,183 +70,26 @@ func (c *CollectHostOS) Collect(progressChan chan<- interface{}) (map[string][]b
 }
 
 func (c *CollectHostOS) RemoteCollect(progressChan chan<- interface{}) (map[string][]byte, error) {
-	allCollectedData := make(map[string][]byte)
-
-	scheme := runtime.NewScheme()
-	if err := corev1.AddToScheme(scheme); err != nil {
-		return nil, errors.Wrap(err, "failed to add runtime scheme")
+	params := &RemoteCollectParams{
+		ProgressChan:  progressChan,
+		HostCollector: &troubleshootv1beta2.HostCollect{HostOS: c.hostCollector},
+		BundlePath:    c.BundlePath,
+		ClientConfig:  c.ClientConfig,
+		Image:         c.Image,
+		PullPolicy:    c.PullPolicy,
+		Timeout:       c.Timeout,
+		LabelSelector: c.LabelSelector,
+		NamePrefix:    c.NamePrefix,
+		Namespace:     c.Namespace,
+		Title:         c.Title(),
 	}
 
-	client, err := kubernetes.NewForConfig(c.ClientConfig)
+	output, err := remoteHostCollect(*params)
 	if err != nil {
-		return nil, err
-	}
-
-	runner := &podRunner{
-		client:       client,
-		scheme:       scheme,
-		image:        c.Image,
-		pullPolicy:   c.PullPolicy,
-		waitInterval: remoteCollectorDefaultInterval,
-	}
-
-	ctx, cancel := context.WithTimeout(context.Background(), c.Timeout)
-	defer cancel()
-
-	// Get all the nodes where we should run.
-	nodes, err := listNodesNamesInSelector(ctx, client, c.LabelSelector)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get the list of nodes matching a nodeSelector")
-	}
-
-	if c.NamePrefix == "" {
-		c.NamePrefix = remoteCollectorNamePrefix
-	}
-
-	// empty host collect var
-	hostCollector := &troubleshootv1beta2.HostCollect{
-		HostOS: c.hostCollector,
-	}
-
-	result, err := runRemote(ctx, runner, nodes, hostCollector, names.SimpleNameGenerator, c.NamePrefix, c.Namespace)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to run collector remotely")
-	}
-
-	for k, v := range result {
-		if curBytes, ok := allCollectedData[k]; ok {
-			var curResults map[string]string
-			if err := json.Unmarshal(curBytes, &curResults); err != nil {
-				progressChan <- errors.Errorf("failed to read existing results for collector %s: %v\n", c.Title(), err)
-				continue
-			}
-			var newResults map[string]string
-			if err := json.Unmarshal(v, &newResults); err != nil {
-				progressChan <- errors.Errorf("failed to read new results for collector %s: %v\n", c.Title(), err)
-				continue
-			}
-			for file, data := range newResults {
-				curResults[file] = data
-			}
-			combinedResults, err := json.Marshal(curResults)
-			if err != nil {
-				progressChan <- errors.Errorf("failed to combine results for collector %s: %v\n", c.Title(), err)
-				continue
-			}
-			allCollectedData[k] = combinedResults
-		} else {
-			allCollectedData[k] = v
-		}
-
-	}
-
-	output := NewResult()
-
-	// save the first result we find in the node and save it
-	for node, result := range allCollectedData {
-		var nodeResult map[string]string
-		if err := json.Unmarshal(result, &nodeResult); err != nil {
-			return nil, errors.Wrap(err, "failed to marshal node results")
-		}
-
-		for _, collectorResult := range nodeResult {
-			var collectedItems HostOSInfo
-			if err := json.Unmarshal([]byte(collectorResult), &collectedItems); err != nil {
-				return nil, errors.Wrap(err, "failed to marshal collector results")
-			}
-
-			b, err := json.MarshalIndent(collectedItems, "", " ")
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to marshal host os info")
-			}
-			nodes = append(nodes, node)
-			output.SaveResult(c.BundlePath, fmt.Sprintf("host-collectors/system/%s/%s", node, HostInfoFileName), bytes.NewBuffer(b))
-		}
-	}
-
-	// check if NODE_LIST_FILE exists
-	_, err = os.Stat(NODE_LIST_FILE)
-	// if it not exists, save the nodes list
-	if err != nil {
-		nodesBytes, err := json.MarshalIndent(HostOSInfoNodes{Nodes: nodes}, "", " ")
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to marshal host os info nodes")
-		}
-		output.SaveResult(c.BundlePath, NODE_LIST_FILE, bytes.NewBuffer(nodesBytes))
+		return nil, errors.Wrap(err, "failed to run remote host os collector")
 	}
 	return output, nil
 }
-
-/*func (c *CollectHostOS) RemoteHostCollect(progressChan chan<- interface{}) (map[string][]byte, error) {
-
-	restConfig, err := k8sutil.GetRESTConfig()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to convert kube flags to rest config")
-	}
-
-	createOpts := CollectorRunOpts{
-		KubernetesRestConfig: restConfig,
-		Image:                "replicated/troubleshoot:latest",
-		Namespace:            "default",
-		Timeout:              defaultTimeout,
-		NamePrefix:           "hostos-remote",
-		ProgressChan:         progressChan,
-	}
-
-	remoteCollector := &troubleshootv1beta2.RemoteCollector{
-		Spec: troubleshootv1beta2.RemoteCollectorSpec{
-			Collectors: []*troubleshootv1beta2.RemoteCollect{
-				{
-					HostOS: &troubleshootv1beta2.RemoteHostOS{},
-				},
-			},
-		},
-	}
-	// empty redactor for now
-	additionalRedactors := &troubleshootv1beta2.Redactor{}
-	// re-use the collect.CollectRemote function to run the remote collector
-	results, err := CollectRemote(remoteCollector, additionalRedactors, createOpts)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to run remote collector")
-	}
-
-	output := NewResult()
-	nodes := []string{}
-
-	// save the first result we find in the node and save it
-	for node, result := range results.AllCollectedData {
-		var nodeResult map[string]string
-		if err := json.Unmarshal(result, &nodeResult); err != nil {
-			return nil, errors.Wrap(err, "failed to marshal node results")
-		}
-
-		for _, collectorResult := range nodeResult {
-			var collectedItems HostOSInfo
-			if err := json.Unmarshal([]byte(collectorResult), &collectedItems); err != nil {
-				return nil, errors.Wrap(err, "failed to marshal collector results")
-			}
-
-			b, err := json.MarshalIndent(collectedItems, "", " ")
-			if err != nil {
-				return nil, errors.Wrap(err, "failed to marshal host os info")
-			}
-			nodes = append(nodes, node)
-			output.SaveResult(c.BundlePath, fmt.Sprintf("host-collectors/system/%s/%s", node, HostInfoFileName), bytes.NewBuffer(b))
-		}
-	}
-
-	// check if NODE_LIST_FILE exists
-	_, err = os.Stat(NODE_LIST_FILE)
-	// if it not exists, save the nodes list
-	if err != nil {
-		nodesBytes, err := json.MarshalIndent(HostOSInfoNodes{Nodes: nodes}, "", " ")
-		if err != nil {
-			return nil, errors.Wrap(err, "failed to marshal host os info nodes")
-		}
-		output.SaveResult(c.BundlePath, NODE_LIST_FILE, bytes.NewBuffer(nodesBytes))
-	}
-	return output, nil
-}*/
 
 /*func (c *CollectHostOS) CheckRBAC(ctx context.Context) error {
 	if _, err := c.IsExcluded(); err != nil {
